pyproject.toml
[tool.poetry]
name = "site_builder"
version = "0.1.0"
description = ""
authors = ["Alex Ledger <services@alexledger.net>"]

[tool.poetry.dependencies]
python = "^3.10"
sqlite-utils = "^3.30"
click = "^8.1.3"
beautifulsoup4 = "^4.11.2"
python-slugify = "^8.0.1"
dateparser = "^1.1.7"
pandoc = "^2.3"
pydantic = "^1.10.6"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.254"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"